refactor: reorganize role mapping utilities and introduce country-specific role display logic
